# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI SecScore — AI Security Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments/{assessmentId}/collaborators:
    get:
      tags: [Collaboration]
      operationId: listCollaborators
      summary: List collaborators for an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Collaborators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collaborator"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    post:
      tags: [Collaboration]
      operationId: addCollaborator
      summary: Share an assessment with another user by email
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorInput"
      responses:
        "201":
          description: Collaborator added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collaborator" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /assessments/{assessmentId}/collaborators/{collaboratorId}:
    delete:
      tags: [Collaboration]
      operationId: removeCollaborator
      summary: Remove a collaborator from an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
        - name: collaboratorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /shared-assessments:
    get:
      tags: [Collaboration]
      operationId: listSharedAssessments
      summary: List assessments shared with the current user
      responses:
        "200":
          description: Shared assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharedAssessment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ── Compliance ──────────────────────────────────────────────────────────────
  /assessments/{assessmentId}/compliance:
    get:
      tags: [Compliance]
      operationId: getComplianceView
      summary: Get standards-mapping compliance view for an assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Compliance view grouped by standard category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComplianceView" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Domain schemas ────────────────────────────────────────────────────────
...
    Collaborator:
      type: object
      required: [id, assessmentId, ownerId, collaboratorEmail, role, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        ownerId:
          type: string
        collaboratorEmail:
          type: string
        role:
          type: string
          enum: [viewer, editor]
        createdAt:
          type: string
          format: date-time
    CollaboratorInput:
      type: object
      required: [collaboratorEmail, role]
      properties:
        collaboratorEmail:
          type: string
          format: email
          minLength: 3
        role:
          type: string
          enum: [viewer, editor]
    SharedAssessment:
      type: object
      required: [id, name, systemName, status, role, ownerEmail]
      properties:
        id:
          type: integer
        name:
          type: string
        systemName:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        role:
          type: string
          enum: [viewer, editor]
        ownerEmail:
          type: ["string", "null"]

    # ── Compliance schemas ────────────────────────────────────────────────────
    ComplianceFramework:
      type: object
      required: [frameworkId, frameworkName, score, riskLevel, answeredCount, totalCount]
      properties:
        frameworkId:
          type: integer
        frameworkName:
          type: string
        referenceUrl:
          type: ["string", "null"]
        score:
          type: number
        riskLevel:
          type: string
          enum: [Critical, High, Medium, Low]
        answeredCount:
          type: integer
        totalCount:
          type: integer
    ComplianceCategory:
      type: object
      required: [category, avgScore, frameworks]
      properties:
        category:
          type: string
        avgScore:
          type: number
        frameworks:
          type: array
          items:
            $ref: "#/components/schemas/ComplianceFramework"
    ComplianceView:
      type: object
      required: [assessmentId, overallScore, categories]
      properties:
        assessmentId:
          type: integer
        overallScore:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ComplianceCategory"