# lib/api-zod/src/generated/types/complianceFramework.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI SecScore — AI Security Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceFrameworkRiskLevel } from './complianceFrameworkRiskLevel';

export interface ComplianceFramework {
  frameworkId: number;
  frameworkName: string;
  /** @nullable */
  referenceUrl?: string | null;
  score: number;
  riskLevel: ComplianceFrameworkRiskLevel;
  answeredCount: number;
  totalCount: number;
}