# lib/api-zod/src/generated/types/complianceView.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI SecScore — AI Security Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceCategory } from './complianceCategory';

export interface ComplianceView {
  assessmentId: number;
  overallScore: number;
  categories: ComplianceCategory[];
}